+++ blueetl/utils.py
@@ -28,7 +28,7 @@ def timed(log: Callable, msg: str, *args) -> Iterator[None]:
         log(f"{msg} [{elapsed:.2f} seconds]", *args)
 
 
-def setup_logging(loglevel: int, logformat: Optional[str] = None, **logparams) -> None:
+def setup_logging(loglevel: Union[int, str], logformat: Optional[str] = None, **logparams) -> None:
     """Setup logging."""
     logformat = logformat or "%(asctime)s %(levelname)s %(name)s: %(message)s"
     logging.basicConfig(format=logformat, level=loglevel, **logparams)