+++ tests/unit/extract/test_simulations.py
@@ -1,8 +1,7 @@
-from unittest.mock import Mock, PropertyMock, patch
+from unittest.mock import Mock, patch
 
 import pandas as pd
 import pytest
-from bluepy.exceptions import BluePyError
 from pandas.testing import assert_frame_equal
 
 from blueetl.extract import simulations as test_module
@@ -24,8 +23,10 @@ def _get_mock_simulation(n=0):
     return mock_simulation
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_config(mock_simulation_class):
+def test_simulations_from_config(mock_simulation_class, mock_is_complete, mock_is_existing):
     mock_simulation0 = _get_mock_simulation()
     mock_simulation1 = _get_mock_simulation()
     mock_circuit0 = mock_simulation0.circuit
@@ -69,10 +70,16 @@ def test_simulations_from_config(mock_simulation_class):
     assert config.to_pandas.call_count == 1
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_config_filtered_by_simulation_id(mock_simulation_class):
+def test_simulations_from_config_filtered_by_simulation_id(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
     mock_simulation0 = _get_mock_simulation()
     mock_simulation1 = _get_mock_simulation()
     mock_circuit0 = mock_simulation0.circuit
@@ -108,10 +115,16 @@ def test_simulations_from_config_filtered_by_simulation_id(mock_simulation_class
     assert config.to_pandas.call_count == 1
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", side_effect=[False, True])
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_config_without_spikes(mock_simulation_class):
+def test_simulations_from_config_without_spikes(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
     mock_simulation0 = _get_mock_simulation()
     mock_simulation1 = _get_mock_simulation()
     mock_circuit0 = mock_simulation0.circuit
@@ -124,8 +137,6 @@ def test_simulations_from_config_without_spikes(mock_simulation_class):
             {"simulation_path": "path2", "seed": 11, "Grating Orientation (degrees)": 45},
         ]
     )
-    # mock the spikes property to simulate an incomplete simulation
-    type(mock_simulation0).spikes = PropertyMock(side_effect=BluePyError)
 
     result = test_module.Simulations.from_config(config)
 
@@ -150,10 +161,65 @@ def test_simulations_from_config_without_spikes(mock_simulation_class):
     assert config.to_pandas.call_count == 1
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", side_effect=[False, True])
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_pandas_load_complete_campaign(mock_simulation_class):
+def test_simulations_from_config_first_nonexistent(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
+    mock_simulation0 = _get_mock_simulation()
+    mock_simulation1 = _get_mock_simulation()
+    mock_circuit0 = mock_simulation0.circuit
+    mock_circuit1 = mock_simulation1.circuit
+    # Simulation is called only once, i.e. when it exists
+    mock_simulation_class.return_value = mock_simulation1
+    config = Mock()
+    config.to_pandas.return_value = pd.DataFrame(
+        [
+            {"simulation_path": "path1", "seed": 10, "Grating Orientation (degrees)": 0},
+            {"simulation_path": "path2", "seed": 11, "Grating Orientation (degrees)": 45},
+        ]
+    )
+
+    result = test_module.Simulations.from_config(config)
+
+    expected_df = pd.DataFrame(
+        [
+            # circuit_id is 1 and not 0, and circuit is mock_circuit1 and not 0,
+            # because the first simulation and circuit cannot be loaded and hashed
+            {
+                "simulation_path": "path2",
+                "seed": 11,
+                "Grating Orientation (degrees)": 45,
+                "simulation_id": 1,
+                "circuit_id": 1,
+                "simulation": mock_simulation1,
+                "circuit": mock_circuit1,
+            },
+        ],
+        index=[1],
+    )
+    expected_df = ensure_dtypes(expected_df)
+    assert isinstance(result, test_module.Simulations)
+    assert_frame_equal(result.df, expected_df)
+
+    assert config.to_pandas.call_count == 1
+    assert mock_simulation_class.call_count == 1
+    assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 1
+
+
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
+@patch(f"{test_module.__name__}.Simulation", autospec=True)
+def test_simulations_from_pandas_load_complete_campaign(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
     mock_simulation0 = _get_mock_simulation()
     mock_simulation1 = _get_mock_simulation()
     mock_circuit0 = mock_simulation0.circuit
@@ -206,10 +272,16 @@ def test_simulations_from_pandas_load_complete_campaign(mock_simulation_class):
 
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_pandas_load_incomplete_campaign(mock_simulation_class):
+def test_simulations_from_pandas_load_incomplete_campaign(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
     mock_simulation0 = _get_mock_simulation()
     mock_simulation1 = _get_mock_simulation()
     mock_circuit0 = mock_simulation0.circuit
@@ -262,10 +334,16 @@ def test_simulations_from_pandas_load_incomplete_campaign(mock_simulation_class)
 
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_pandas_load_inconsistent_campaign(mock_simulation_class):
+def test_simulations_from_pandas_load_inconsistent_campaign(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
     # the mocked simulations have different circuit_hash,
     # the dataframe simulations have the same circuit_id
     mock_simulation0 = _get_mock_simulation(0)
@@ -296,10 +374,54 @@ def test_simulations_from_pandas_load_inconsistent_campaign(mock_simulation_clas
 
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
+@patch(f"{test_module.__name__}._is_simulation_existing", side_effect=[False, True])
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
 @patch(f"{test_module.__name__}.Simulation", autospec=True)
-def test_simulations_from_pandas_filtered_by_simulation_id(mock_simulation_class):
+def test_simulations_from_pandas_first_nonexistent(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
+    mock_simulation0 = _get_mock_simulation()
+    mock_simulation1 = _get_mock_simulation()
+    mock_circuit0 = mock_simulation0.circuit
+    mock_circuit1 = mock_simulation1.circuit
+    mock_simulation_class.side_effect = [mock_simulation0, mock_simulation1]
+    df = pd.DataFrame(
+        [
+            {
+                "simulation_path": "path1",
+                "seed": 10,
+                "Grating Orientation (degrees)": 0,
+                "simulation_id": 0,
+                "circuit_id": 0,
+            },
+            {
+                "simulation_path": "path2",
+                "seed": 11,
+                "Grating Orientation (degrees)": 45,
+                "simulation_id": 1,
+                "circuit_id": 0,
+            },
+        ]
+    )
+    with pytest.raises(RuntimeError, match="Inconsistent cached simulations"):
+        test_module.Simulations.from_pandas(df)
+
+    assert mock_simulation_class.call_count == 1
+    assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 1
+
+
+@patch(f"{test_module.__name__}._is_simulation_existing", return_value=True)
+@patch(f"{test_module.__name__}._is_simulation_complete", return_value=True)
+@patch(f"{test_module.__name__}.Simulation", autospec=True)
+def test_simulations_from_pandas_filtered_by_simulation_id(
+    mock_simulation_class, mock_is_complete, mock_is_existing
+):
     mock_simulation0 = _get_mock_simulation()
     mock_simulation1 = _get_mock_simulation()
     mock_circuit0 = mock_simulation0.circuit
@@ -346,6 +468,8 @@ def test_simulations_from_pandas_filtered_by_simulation_id(mock_simulation_class
 
     assert mock_simulation_class.call_count == 2
     assert mock_circuit0 != mock_circuit1
+    assert mock_is_existing.call_count == 2
+    assert mock_is_complete.call_count == 2
 
 
 def test_simulations_to_pandas():